flask-server/requirements/dev.txt
alembic==1.5.2
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.22
unicode-slugify==0.1.3
Unidecode==1.1.2
Werkzeug==1.0.1
